refactor: extract read path into dev/read.rs
Move the read/mapping path (get_mapping, get_l2_entry/entries, the
do_read_* helpers and read_at) into a new dev/read.rs submodule as an
impl extension of Qcow2Dev. do_read_compressed and get_l2_entry are
widened to pub(crate) since the COW write path and checker call them.
No logic changes.
